bugfix and new features
1.0 (6) BugFix and new features.
-----------
New features to test ->
- Complete integration of mapKit.
- User location request does now work
- Data persistance
- Cloud sync

- Test UI/UX and suggest new feature you might want to add

When the app is launched for the first time it now asks for user location (Allow once, Allow when in use, don't allow).
When creating a new match, is now possibile to select a place using the power of Apple Maps itself! It is possibile to select a pitch or open the exact place on maps, in addition of look around mode.
Start position is no longer fixed.

N.B. Information will not be stored at all, that means that the application is not working as it is intended at the moment